+++ beacon-chain/deprecated-blockchain/fork_choice_deprecated.go
