+++ beacon-chain/deprecated-blockchain/service.go
@@ -4,6 +4,7 @@
 package blockchain
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"runtime"
@@ -11,14 +12,15 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
+	"github.com/prysmaticlabs/go-ssz"
 	"github.com/prysmaticlabs/prysm/beacon-chain/attestation"
-	"github.com/prysmaticlabs/prysm/beacon-chain/blockchain/forkchoice"
 	"github.com/prysmaticlabs/prysm/beacon-chain/cache/depositcache"
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/state"
+	b "github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations"
 	"github.com/prysmaticlabs/prysm/beacon-chain/p2p"
 	"github.com/prysmaticlabs/prysm/beacon-chain/powchain"
+	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
 	"github.com/prysmaticlabs/prysm/shared/event"
 	"github.com/sirupsen/logrus"
@@ -41,14 +43,17 @@ type ChainService struct {
 	beaconDB             db.Database
 	depositCache         *depositcache.DepositCache
 	web3Service          *powchain.Web3Service
+	attsService          attestation.TargetHandler
 	opsPoolService       operations.OperationFeeds
-	forkChoiceStore      *forkchoice.Store
 	chainStartChan       chan time.Time
+	canonicalBlockFeed   *event.Feed
 	genesisTime          time.Time
+	finalizedEpoch       uint64
 	stateInitializedFeed *event.Feed
 	p2p                  p2p.Broadcaster
 	canonicalRoots       map[uint64][]byte
 	canonicalRootsLock   sync.RWMutex
+	receiveBlockLock     sync.Mutex
 	maxRoutines          int64
 	headSlot             uint64
 }
@@ -61,6 +66,7 @@ type Config struct {
 	BeaconDB       db.Database
 	DepositCache   *depositcache.DepositCache
 	OpsPoolService operations.OperationFeeds
+	DevMode        bool
 	P2p            p2p.Broadcaster
 	MaxRoutines    int64
 }
@@ -69,7 +75,6 @@ type Config struct {
 // be registered into a running beacon node.
 func NewChainService(ctx context.Context, cfg *Config) (*ChainService, error) {
 	ctx, cancel := context.WithCancel(ctx)
-	store := forkchoice.NewForkChoiceService(ctx, cfg.BeaconDB)
 	return &ChainService{
 		ctx:                  ctx,
 		cancel:               cancel,
@@ -77,7 +82,8 @@ func NewChainService(ctx context.Context, cfg *Config) (*ChainService, error) {
 		depositCache:         cfg.DepositCache,
 		web3Service:          cfg.Web3Service,
 		opsPoolService:       cfg.OpsPoolService,
-		forkChoiceStore:      store,
+		attsService:          cfg.AttsService,
+		canonicalBlockFeed:   new(event.Feed),
 		chainStartChan:       make(chan time.Time),
 		stateInitializedFeed: new(event.Feed),
 		p2p:                  cfg.P2p,
@@ -96,6 +102,7 @@ func (c *ChainService) Start() {
 	if beaconState != nil {
 		log.Info("Beacon chain data already exists, starting service")
 		c.genesisTime = time.Unix(int64(beaconState.GenesisTime), 0)
+		c.finalizedEpoch = beaconState.FinalizedCheckpoint.Epoch
 	} else {
 		log.Info("Waiting for ChainStart log from the Validator Deposit Contract to start the beacon chain...")
 		if c.web3Service == nil {
@@ -115,9 +122,11 @@ func (c *ChainService) Start() {
 // deposit contract, initializes the beacon chain's state, and kicks off the beacon chain.
 func (c *ChainService) processChainStartTime(genesisTime time.Time, chainStartSub event.Subscription) {
 	initialDeposits := c.web3Service.ChainStartDeposits()
-	if err := c.initializeBeaconChain(genesisTime, initialDeposits, c.web3Service.ChainStartETH1Data()); err != nil {
+	beaconState, err := c.initializeBeaconChain(genesisTime, initialDeposits, c.web3Service.ChainStartETH1Data())
+	if err != nil {
 		log.Fatalf("Could not initialize beacon chain: %v", err)
 	}
+	c.finalizedEpoch = beaconState.FinalizedCheckpoint.Epoch
 	c.stateInitializedFeed.Send(genesisTime)
 	chainStartSub.Unsubscribe()
 }
@@ -125,25 +134,63 @@ func (c *ChainService) processChainStartTime(genesisTime time.Time, chainStartSu
 // initializes the state and genesis block of the beacon chain to persistent storage
 // based on a genesis timestamp value obtained from the ChainStart event emitted
 // by the ETH1.0 Deposit Contract and the POWChain service of the node.
-func (c *ChainService) initializeBeaconChain(genesisTime time.Time, deposits []*ethpb.Deposit, eth1data *ethpb.Eth1Data) error {
-	_, span := trace.StartSpan(context.Background(), "beacon-chain.ChainService.initializeBeaconChain")
+func (c *ChainService) initializeBeaconChain(genesisTime time.Time, deposits []*ethpb.Deposit, eth1data *ethpb.Eth1Data) (*pb.BeaconState, error) {
+	ctx, span := trace.StartSpan(context.Background(), "beacon-chain.ChainService.initializeBeaconChain")
 	defer span.End()
 	log.Info("ChainStart time reached, starting the beacon chain!")
 	c.genesisTime = genesisTime
 	unixTime := uint64(genesisTime.Unix())
+	// TODO(3219): Use new fork choice service.
+	db, isLegacyDB := c.beaconDB.(*db.BeaconDB)
+	if !isLegacyDB {
+		panic("Chain service cannot operate with the new database interface due to in-progress fork choice changes")
+	}
 
-	beaconState, err := state.GenesisBeaconState(deposits, unixTime, eth1data)
+	if err := db.InitializeState(ctx, unixTime, deposits, eth1data); err != nil {
+		return nil, errors.Wrap(err, "could not initialize beacon state to disk")
+	}
+	beaconState, err := c.beaconDB.HeadState(ctx)
 	if err != nil {
-		return errors.Wrap(err, "could not initialize genesis state")
+		return nil, errors.Wrap(err, "could not attempt fetch beacon state")
 	}
 
-	if err := c.forkChoiceStore.GenesisStore(c.ctx, beaconState); err != nil {
-		return errors.Wrap(err, "could not start gensis store for fork choice")
+	stateRoot, err := ssz.HashTreeRoot(beaconState)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not hash beacon state")
+	}
+	genBlock := b.NewGenesisBlock(stateRoot[:])
+	genBlockRoot, err := ssz.SigningRoot(genBlock)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not hash beacon block")
 	}
 
-	c.canonicalRoots[beaconState.Slot] = c.FinalizedCheckpt().Root
+	if err := c.beaconDB.SaveBlock(ctx, genBlock); err != nil {
+		return nil, errors.Wrap(err, "could not save genesis block to disk")
+	}
 
-	return nil
+	if err := db.SaveAttestationTarget(ctx, &pb.AttestationTarget{
+		Slot:            genBlock.Slot,
+		BeaconBlockRoot: genBlockRoot[:],
+		ParentRoot:      genBlock.ParentRoot,
+	}); err != nil {
+		return nil, errors.Wrap(err, "failed to save attestation target")
+	}
+	if err := db.UpdateChainHead(ctx, genBlock, beaconState); err != nil {
+		return nil, errors.Wrap(err, "could not set chain head")
+	}
+	if err := db.SaveJustifiedBlock(genBlock); err != nil {
+		return nil, errors.Wrap(err, "could not save genesis block as justified block")
+	}
+	if err := db.SaveFinalizedBlock(genBlock); err != nil {
+		return nil, errors.Wrap(err, "could not save genesis block as finalized block")
+	}
+	if err := db.SaveJustifiedState(beaconState); err != nil {
+		return nil, errors.Wrap(err, "could not save genesis state as justified state")
+	}
+	if err := db.SaveFinalizedState(beaconState); err != nil {
+		return nil, errors.Wrap(err, "could not save genesis state as finalized state")
+	}
+	return beaconState, nil
 }
 
 // Stop the blockchain service's main event loop and associated goroutines.
@@ -163,8 +210,47 @@ func (c *ChainService) Status() error {
 	return nil
 }
 
+// CanonicalBlockFeed returns a channel that is written to
+// whenever a new block is determined to be canonical in the chain.
+func (c *ChainService) CanonicalBlockFeed() *event.Feed {
+	return c.canonicalBlockFeed
+}
+
 // StateInitializedFeed returns a feed that is written to
 // when the beacon state is first initialized.
 func (c *ChainService) StateInitializedFeed() *event.Feed {
 	return c.stateInitializedFeed
+}
+
+// ChainHeadRoot returns the hash root of the last beacon block processed by the
+// block chain service.
+func (c *ChainService) ChainHeadRoot(ctx context.Context) ([32]byte, error) {
+	head, err := c.beaconDB.HeadBlock(ctx)
+	if err != nil {
+		return [32]byte{}, errors.Wrap(err, "could not retrieve chain head")
+	}
+
+	root, err := ssz.SigningRoot(head)
+	if err != nil {
+		return [32]byte{}, errors.Wrap(err, "could not tree hash parent block")
+	}
+	return root, nil
+}
+
+// UpdateCanonicalRoots sets a new head into the canonical block roots map.
+func (c *ChainService) UpdateCanonicalRoots(newHead *ethpb.BeaconBlock, newHeadRoot [32]byte) {
+	c.canonicalRootsLock.Lock()
+	defer c.canonicalRootsLock.Unlock()
+	c.canonicalRoots[newHead.Slot] = newHeadRoot[:]
+}
+
+// IsCanonical returns true if the input block hash of the corresponding slot
+// is part of the canonical chain. False otherwise.
+func (c *ChainService) IsCanonical(slot uint64, hash []byte) bool {
+	c.canonicalRootsLock.RLock()
+	defer c.canonicalRootsLock.RUnlock()
+	if canonicalHash, ok := c.canonicalRoots[slot]; ok {
+		return bytes.Equal(canonicalHash, hash)
+	}
+	return false
 }
